+++ core/modules/modEmergencyHouse.class.php
@@ -70,6 +70,7 @@ class modEmergencyHouse extends DolibarrModules
 			),
 			'substitutions' => 1,
 			'models' => 1,
+			'css' => array('/emergencyhouse/css/emergencyhouse.css.php'),
 		);
 		$this->dirs = array('/emergencyhouse/temp');
 		$this->tabs = array();
@@ -235,7 +236,7 @@ class modEmergencyHouse extends DolibarrModules
 			'fk_menu' => '',
 			'type' => 'top',
 			'titre' => 'EmergencyHouse',
-			'prefix' => img_picto('', $this->picto, 'class="pictofixedwidth valignmiddle"'),
+			'prefix' => img_picto('', $this->picto, 'class="paddingright pictofixedwidth valignmiddle"'),
 			'mainmenu' => 'emergencyhouse',
 			'leftmenu' => '',
 			'url' => '/emergencyhouse/index.php',